refactor(compare): extract I/O helpers to report_io.py (Phase A 1/N)
Move atomic file writes (write_text_atomic, write_json_atomic_strict) and
the thresholds.yaml loader into a new hv_dataqc/compare/report_io.py.
Pure mechanical extraction: no behavior change. Functions renamed from
leading-underscore to public form since they're now module-public.

Renamed `_write_json_atomic` → `write_json_atomic_strict` (the _strict
suffix distinguishes it from hv_dataqc_common.write_json_atomic, which it
wraps with stricter encoding settings).

Plan called this `io.py` but renamed to `report_io.py` to avoid the
stdlib `io` module name collision — fully-qualified imports work but
`from io import ...` would be ambiguous.

Regression-check passes on COPDGene, ARIC, HCHS (MD + JSON both match
baseline). Unit tests: 79/79 pass.
